tests/wpt/web-platform-tests/html/webappapis/animation-frames/callback-timestamp.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AnimationTiming Test: FrameRequestCallback - timestamp argument</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#animation-frames">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

  async_test(t => {
    requestAnimationFrame(t.step_func_done(time => {
      assert_equals(typeof time, "number", "callback contains a number argument");
    }))
  }, "Check FrameRequestCallback has a DOMHighResTimeStamp argument");

</script>
